move prediction exportation within ligntning module
