--- configs/export/default.yaml
@@ -1,7 +1,0 @@
-_target_: src.export.defile_export.DefileExport
-
-species: ${data.species}
-dirpath: ${paths.output_dir}
-filename: test_pred.nc
-plot: true
-plotdir: ${paths.output_dir}/figures
